refactor: ui-centric code to own package
