+++ openai-connector-assistant/plugin.xml
@@ -10,13 +10,13 @@
                id="chatGpt.question"
                name="question"
                optional="true"
-               values="com.axonivy.connector.openai.assistant.ChatGPTquest">
+               values="com.axonivy.connector.openai.assistant.ui.ChatGPTquest">
          </commandParameter>
       </command>
    </extension>
    <extension point="org.eclipse.ui.handlers">
       <handler commandId="openai.commands.assist"
-        class="com.axonivy.connector.openai.assistant.ChatGPTAssistantHandler">
+        class="com.axonivy.connector.openai.assistant.ui.ChatGPTAssistantHandler">
       </handler>
    </extension>
    <extension point="org.eclipse.ui.menus">
